+++ exercicio1/src/main/java/br/com/andre/exercicio1/MultiplesOfAll.java
@@ -4,9 +4,9 @@ import java.util.Set;
 
 public class MultiplesOfAll implements MultiplesCalculationStrategy {
     @Override
-    public boolean isMultiple(int n, Set<Integer> baseOperands) {
+    public boolean isMultiple(long n, Set<Long> baseOperands) {
         var result = true;
-        for (int operand : baseOperands) {
+        for (long operand : baseOperands) {
             result &= n % operand == 0;
         }
         return result;
