+++ exercicio1/src/main/java/br/com/andre/exercicio1/MultiplesOfAny.java
@@ -4,7 +4,7 @@ import java.util.Set;
 
 public class MultiplesOfAny implements MultiplesCalculationStrategy {
     @Override
-    public boolean isMultiple(int n, Set<Integer> baseOperands) {
+    public boolean isMultiple(long n, Set<Long> baseOperands) {
         return baseOperands.stream().map(operand -> {
             return n % operand == 0;
         }).reduce((previousIsMultiple, result) -> {
